+++ src/database.js
@@ -1,0 +1,9 @@
+const mongoose = require("mongoose");
+
+mongoose.set("strictQuery", false);
+mongoose
+  .connect(process.env.MONGODB_URI)
+  .then(() => console.log("Connected to MongoDB Atlas"))
+  .catch((error) => console.error(error));
+
+module.exports = mongoose;
